API problems resolved / Adding and Deleting functionality fully working

+++ app/src/main/java/com/example/fyp/data/repository/RoomRepository.kt
@@ -10,7 +10,7 @@ constructor(private val receiptDAO: ReceiptDAO) {
     fun getAll(): Flow<List<ReceiptModel>>
             = receiptDAO.getAll()
 
-    fun get(id: Int) = receiptDAO.get(id)
+    fun get(id: Long) = receiptDAO.get(id)
 
     suspend fun insert(receipt: ReceiptModel)
     { receiptDAO.insert(receipt) }
